refactor: Shuffling around of persistable types
