+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/sweep/asts/SweepableBucket.java
@@ -29,23 +29,4 @@ public interface SweepableBucket {
     static SweepableBucket of(Bucket bucket, TimestampRange timestampRange) {
         return ImmutableSweepableBucket.of(bucket, timestampRange);
     }
-
-    @Value.Immutable
-    interface TimestampRange {
-        @Value.Parameter
-        long startInclusive();
-
-        @Value.Parameter
-        long endExclusive();
-
-        // TODO(mdaudali): Should this be named closeBucket? It's really just a range, so it feels like no.
-        static TimestampRange of(long startInclusive, long endExclusive) {
-            return ImmutableTimestampRange.of(startInclusive, endExclusive);
-        }
-
-        static TimestampRange openBucket(long startInclusive) {
-            // Think very carefully about changing from -1 without a migration
-            return of(startInclusive, -1);
-        }
-    }
 }
